remove casting warnings
git-svn-id: https://svn.aspect.dealii.org/trunk/aspect@660 244a8438-9d6d-46c2-a098-b74c898f6b90

+++ source/material_model/steinberger.cc
@@ -64,7 +64,7 @@ namespace aspect
 
             Assert(depth>=min_depth, ExcMessage("not in range"));
             Assert(depth<=max_depth, ExcMessage("not in range"));
-            const unsigned int idx = (depth-min_depth)/delta_depth;
+            const unsigned int idx = static_cast<unsigned int>((depth-min_depth)/delta_depth);
             Assert(idx<values.size(), ExcMessage("not in range"));
             return values[idx];
           }
@@ -116,7 +116,7 @@ namespace aspect
 
             Assert(depth>=min_depth, ExcMessage("not in range"));
             Assert(depth<=max_depth, ExcMessage("not in range"));
-            const unsigned int idx = (depth-min_depth)/delta_depth;
+            const unsigned int idx = static_cast<unsigned int>((depth-min_depth)/delta_depth);
             Assert(idx<values.size(), ExcMessage("not in range"));
             return values[idx];
           }
